+++ supabase/functions/_shared/scheduledcron/cron.ts
@@ -1,3 +1,5 @@
+import { dateToCron } from './helpers.ts'
+
 const invokeBroadcastCron = (runAt: number | Date): string => {
   const runTime = dateToCron(new Date(runAt))
   return `
@@ -113,49 +115,10 @@ const handleFailedDeliveriesCron = (): string => {
   `
 }
 
-const UNSCHEDULE_INVOKE_BROADCAST = "SELECT cron.unschedule('invoke-broadcast');"
-const UNSCHEDULE_SEND_FIRST_MESSAGES = "SELECT cron.unschedule('send-first-messages');"
-const UNSCHEDULE_SEND_SECOND_MESSAGES = "SELECT cron.unschedule('send-second-messages');"
-const UNSCHEDULE_DELAY_SEND_SECOND_MESSAGES = "SELECT cron.unschedule('delay-send-second-messages');"
-const UNSCHEDULE_RECONCILE_TWILLIO = "SELECT cron.unschedule('reconcile-twilio-status');"
-const UNSCHEDULE_DELAY_RECONCILE_TWILLIO = "SELECT cron.unschedule('delay-reconcile-twilio-status');"
-const UNSCHEDULE_HANDLE_FAILED_DELIVERIES = "SELECT cron.unschedule('handle-failed-deliveries');"
-const SELECT_JOB_NAMES = 'SELECT jobname from cron.job;'
-
-const JOB_NAMES = [
-  'invoke-broadcast',
-  'send-first-messages',
-  'send-second-messages',
-  'delay-send-second-messages',
-  'reconcile-twilio-status',
-  'delay-reconcile-twilio-status',
-  'handle-failed-deliveries',
-]
-
-const dateToCron = (date: Date) => {
-  const minutes = date.getMinutes()
-  const hours = date.getHours()
-  const days = date.getDate()
-  const months = date.getMonth() + 1
-  const dayOfWeek = date.getDay()
-
-  return `${minutes} ${hours} ${days} ${months} ${dayOfWeek}`
-}
-
 export {
-  dateToCron,
   handleFailedDeliveriesCron,
   invokeBroadcastCron,
-  JOB_NAMES,
   reconcileTwilioStatusCron,
-  SELECT_JOB_NAMES,
   sendFirstMessagesCron,
   sendSecondMessagesCron,
-  UNSCHEDULE_DELAY_RECONCILE_TWILLIO,
-  UNSCHEDULE_DELAY_SEND_SECOND_MESSAGES,
-  UNSCHEDULE_HANDLE_FAILED_DELIVERIES,
-  UNSCHEDULE_INVOKE_BROADCAST,
-  UNSCHEDULE_RECONCILE_TWILLIO,
-  UNSCHEDULE_SEND_FIRST_MESSAGES,
-  UNSCHEDULE_SEND_SECOND_MESSAGES,
 }
