+++ supabase/functions/_shared/services/BroadcastService.ts
@@ -1,12 +1,9 @@
 import { and, eq, inArray, isNull, lt, or, sql } from 'drizzle-orm'
-import { PostgresJsTransaction } from 'drizzle-orm/postgres-js'
 
 import supabase from '../lib/supabase.ts'
 import DateUtils from '../misc/DateUtils.ts'
 import {
-  AudienceSegment,
   authors,
-  Broadcast,
   broadcasts,
   BroadcastSegment,
   broadcastSentMessageStatus,
@@ -17,49 +14,36 @@ import {
 } from '../drizzle/schema.ts'
 import {
   handleFailedDeliveriesCron,
-  invokeBroadcastCron,
   reconcileTwilioStatusCron,
   sendFirstMessagesCron,
   sendSecondMessagesCron,
-  UNSCHEDULE_DELAY_RECONCILE_TWILLIO,
-  UNSCHEDULE_DELAY_SEND_SECOND_MESSAGES,
-  UNSCHEDULE_HANDLE_FAILED_DELIVERIES,
-  UNSCHEDULE_INVOKE_BROADCAST,
-  UNSCHEDULE_RECONCILE_TWILLIO,
-  UNSCHEDULE_SEND_FIRST_MESSAGES,
-  UNSCHEDULE_SEND_SECOND_MESSAGES,
 } from '../scheduledcron/cron.ts'
-import {
-  BroadcastResponse,
-  BroadcastUpdate,
-  cloneBroadcast,
-  convertToBroadcastMessagesStatus,
-  convertToFutureBroadcast,
-  convertToPastBroadcast,
-  convertToUpcomingBroadcast,
-  UpcomingBroadcastResponse,
-} from '../dto/BroadcastRequestResponse.ts'
+import { cloneBroadcast, convertToBroadcastMessagesStatus } from '../dto/BroadcastRequestResponse.ts'
 import MissiveUtils from '../lib/Missive.ts'
 import { sleep } from '../misc/utils.ts'
 import {
-  BroadcastDashBoardQueryReturn,
   FAILED_DELIVERED_QUERY,
   insertOutgoingMessagesFallbackQuery,
-  insertOutgoingMessagesQuery,
-  selectBroadcastDashboard,
+  queueBroadcastMessages,
+  UNSCHEDULE_COMMANDS,
 } from '../scheduledcron/queries.ts'
 import NotFoundError from '../exception/NotFoundError.ts'
-import { MISSIVE_API_RATE_LIMIT } from '../constants/constants.ts'
 import BadRequestError from '../exception/BadRequestError.ts'
-import { isBroadcastRunning } from '../scheduledcron/helpers.ts'
 import Sentry from '../lib/Sentry.ts'
+import {
+  insertBroadcastSegmentRecipients,
+  isBroadcastRunning,
+  makeNextBroadcastSchedule,
+} from './BroadcastServiceUtils.ts'
+import { MISSIVE_API_RATE_LIMIT } from '../constants.ts'
 
 const makeBroadcast = async (): Promise<void> => {
   const isRunning = await isBroadcastRunning()
   if (isRunning) {
     throw new BadRequestError('Unable to make broadcast: another broadcast is running')
   }
-  const nextBroadcast = await supabase.query.broadcasts.findFirst({
+
+  const broadcast = await supabase.query.broadcasts.findFirst({
     where: and(
       eq(broadcasts.editable, true),
       lt(broadcasts.runAt, DateUtils.advance(24 * 60 * 60 * 1000)), // TODO: Prevent making 2 broadcast in a day
@@ -72,27 +56,17 @@ const makeBroadcast = async (): Promise<void> => {
       },
     },
   })
-  if (!nextBroadcast) {
-    throw new NotFoundError('Next broadcast not found')
-  }
-  if (nextBroadcast.broadcastToSegments.length === 0) {
-    console.log(`Broadcast has no associated segment. Data: ${JSON.stringify(nextBroadcast)}`)
-    throw new Error('Broadcast has no associated segment.')
+  if (!broadcast) {
+    throw new NotFoundError('Broadcast not found')
   }
 
   await supabase.transaction(async (tx) => {
-    for (const broadcastSegment of nextBroadcast.broadcastToSegments) {
-      await insertBroadcastSegmentRecipients(tx, broadcastSegment, nextBroadcast)
-    }
-    const fallbackStatement = await insertOutgoingMessagesFallbackQuery(tx, nextBroadcast)
-    if (fallbackStatement) {
-      await tx.execute(sql.raw(fallbackStatement))
-    }
-    await makeNextBroadcastSchedule(tx, nextBroadcast)
-    await tx.execute(sql.raw(sendFirstMessagesCron(nextBroadcast.id)))
-    await tx.update(broadcasts).set({ editable: false }).where(eq(broadcasts.id, nextBroadcast.id))
+    await tx.execute(sql.raw(queueBroadcastMessages(broadcast.id)))
+    await makeNextBroadcastSchedule(tx, broadcast)
+    await tx.execute(sql.raw(sendFirstMessagesCron(broadcast.id)))
+    await tx.update(broadcasts).set({ editable: false }).where(eq(broadcasts.id, broadcast.id))
   })
-  await supabase.execute(sql.raw(UNSCHEDULE_INVOKE_BROADCAST))
+  await supabase.execute(sql.raw(UNSCHEDULE_COMMANDS.INVOKE_BROADCAST))
 }
 
 const sendNow = async (): Promise<void> => {
@@ -147,7 +121,7 @@ const sendNow = async (): Promise<void> => {
 }
 
 const sendBroadcastSecondMessages = async (broadcastID: number) => {
-  // This function is called by a CRON job every minute
+
   const startTime = Date.now()
   const results = await supabase
     .select()
@@ -164,8 +138,8 @@ const sendBroadcastSecondMessages = async (broadcastID: number) => {
 
   if (results.length === 0) {
     // No messages found, unschedule CRON jobs
-    await supabase.execute(sql.raw(UNSCHEDULE_DELAY_SEND_SECOND_MESSAGES))
-    await supabase.execute(sql.raw(UNSCHEDULE_SEND_SECOND_MESSAGES))
+    await supabase.execute(sql.raw(UNSCHEDULE_COMMANDS.DELAY_SEND_SECOND_MESSAGES))
+    await supabase.execute(sql.raw(UNSCHEDULE_COMMANDS.SEND_SECOND_MESSAGES))
     await supabase.execute(sql.raw(reconcileTwilioStatusCron(broadcastID)))
     return
   }
@@ -220,7 +194,13 @@ const sendBroadcastSecondMessages = async (broadcastID: number) => {
 }
 
 const sendBroadcastFirstMessages = async (broadcastID: number) => {
-  // This function is called by a CRON job every minute
+  const result = await supabase.execute(sql`
+    SELECT pgmq.send(
+      'foo',
+      ${JSON.stringify({ hello: 'world' })}::jsonb,
+      ${30}
+    )
+  `)
   const startTime = Date.now()
   const results = await supabase
     .select()
@@ -239,7 +219,7 @@ const sendBroadcastFirstMessages = async (broadcastID: number) => {
     // No messages found, finished sending first messages
     await supabase.execute(sql.raw(sendSecondMessagesCron(startTime, broadcastID, 1)))
     // TODO: There may be some messages that were not processed in the last batch
-    await supabase.execute(sql.raw(UNSCHEDULE_SEND_FIRST_MESSAGES))
+    await supabase.execute(sql.raw(UNSCHEDULE_COMMANDS.SEND_FIRST_MESSAGES))
     return
   }
   const idsToMarkAsProcessed = results.map((outgoing: OutgoingMessage) => outgoing.id!)
@@ -292,60 +272,6 @@ const sendBroadcastFirstMessages = async (broadcastID: number) => {
   }
 }
 
-const getAll = async (
-  limit = 5, // Limit past batches
-  cursor?: number,
-): Promise<BroadcastResponse> => {
-  const selectQuery = selectBroadcastDashboard(cursor ? limit : limit + 1, cursor)
-  const results: BroadcastDashBoardQueryReturn[] = await supabase.execute(sql.raw(selectQuery))
-  // "runAt" value should be a date, but it appears as a string when used in Supabase.
-  results.forEach((broadcast) => broadcast.runAt = new Date(broadcast.runAt))
-  const response = new BroadcastResponse()
-  if (results.length === 0) {
-    return response
-  }
-  if (cursor && results[0].runAt < new Date()) {
-    response.past = results.map((broadcast) => convertToPastBroadcast(broadcast))
-  } else {
-    response.upcoming = convertToUpcomingBroadcast(results[0])
-    response.past = results.slice(1).map((broadcast) => convertToPastBroadcast(broadcast))
-  }
-
-  const lastRunAtTimestamp = results[results.length - 1].runAt.getTime() / 1000
-  response.currentCursor = Math.max(Math.floor(lastRunAtTimestamp) - 1, 0)
-  return response
-}
-
-const patch = async (
-  id: number,
-  broadcast: BroadcastUpdate,
-): Promise<UpcomingBroadcastResponse | undefined> => {
-  return await supabase.transaction(async (tx) => {
-    const result: Broadcast[] = await tx.update(broadcasts)
-      .set({
-        firstMessage: broadcast.firstMessage,
-        secondMessage: broadcast.secondMessage,
-        runAt: broadcast.runAt ? new Date(broadcast.runAt * 1000) : undefined,
-        delay: broadcast.delay,
-      })
-      .where(and(eq(broadcasts.id, id), eq(broadcasts.editable, true)))
-      .returning()
-    if (result.length === 0) {
-      return
-    }
-    if (broadcast.runAt) {
-      try {
-        await tx.execute(sql.raw(UNSCHEDULE_INVOKE_BROADCAST))
-      } catch (e) {
-        console.error(`Failed to unschedule invoke-broadcast: ${e}`)
-      }
-      const invokeNextBroadcast = invokeBroadcastCron(broadcast.runAt * 1000)
-      await tx.execute(sql.raw(invokeNextBroadcast))
-    }
-    return convertToUpcomingBroadcast(result[0])
-  })
-}
-
 const reconcileTwilioStatus = async (broadcastId: number) => {
   const CHUNK_SIZE = 200
   const MAX_RETRIES = 3
@@ -367,8 +293,8 @@ const reconcileTwilioStatus = async (broadcastId: number) => {
       .limit(CHUNK_SIZE)
       .execute()
     if (failedStatusConversations.length === 0) {
-      await supabase.execute(sql.raw(UNSCHEDULE_RECONCILE_TWILLIO))
-      await supabase.execute(sql.raw(UNSCHEDULE_DELAY_RECONCILE_TWILLIO))
+      await supabase.execute(sql.raw(UNSCHEDULE_COMMANDS.RECONCILE_TWILIO))
+      await supabase.execute(sql.raw(UNSCHEDULE_COMMANDS.DELAY_RECONCILE_TWILIO))
       await supabase.execute(sql.raw(handleFailedDeliveriesCron()))
       return
     }
@@ -431,7 +357,7 @@ const handleFailedDeliveries = async () => {
   const failedDelivers = await supabase.execute(sql.raw(FAILED_DELIVERED_QUERY))
   if (!failedDelivers || failedDelivers.length === 0) {
     try {
-      await supabase.execute(sql.raw(UNSCHEDULE_HANDLE_FAILED_DELIVERIES))
+      await supabase.execute(sql.raw(UNSCHEDULE_COMMANDS.HANDLE_FAILED_DELIVERIES))
     } catch (e) {
       console.error(`Failed to unschedule handleFailedDeliveries: ${e}`)
     }
@@ -511,69 +437,11 @@ const handleFailedDeliveries = async () => {
   }
 }
 
-/* ======================================== UTILS ======================================== */
-
-const makeNextBroadcastSchedule = async (
-  // deno-lint-ignore no-explicit-any
-  tx: PostgresJsTransaction<any, any>,
-  previousBroadcast: Broadcast & { broadcastToSegments: { segment: AudienceSegment; ratio: number }[] },
-): Promise<void> => {
-  const newBroadcast = convertToFutureBroadcast(previousBroadcast)
-  const invokeNextBroadcast = invokeBroadcastCron(newBroadcast.runAt)
-  await tx.execute(sql.raw(invokeNextBroadcast))
-  const insertedIds: {
-    id: number
-  }[] = await tx.insert(broadcasts).values(newBroadcast).returning({ id: broadcasts.id })
-  await tx.insert(broadcastsSegments).values(previousBroadcast.broadcastToSegments.map((broadcastSegment) => ({
-    broadcastId: insertedIds[0].id!,
-    segmentId: broadcastSegment.segment.id!,
-    ratio: broadcastSegment.ratio,
-  })))
-}
-
-const insertBroadcastSegmentRecipients = async (
-  // deno-lint-ignore no-explicit-any
-  tx: PostgresJsTransaction<any, any>,
-  broadcastSegment: BroadcastSegment,
-  nextBroadcast: Broadcast,
-) => {
-  // every user receives 2 messages
-  const limit = Math.floor(broadcastSegment.ratio * nextBroadcast.noUsers! / 100)
-  const statement = insertOutgoingMessagesQuery(broadcastSegment, nextBroadcast, limit)
-  await tx.execute(sql.raw(statement))
-}
-
-const updateSubscriptionStatus = async (
-  conversationId: string,
-  phoneNumber: string,
-  isUnsubscribe: boolean,
-  authorName?: string,
-) => {
-  await supabase
-    .update(authors)
-    .set({ unsubscribed: isUnsubscribe })
-    .where(eq(authors.phoneNumber, phoneNumber))
-
-  const action = isUnsubscribe ? 'unsubscribed' : 'resubscribed'
-  const byAuthor = authorName ? ` by ${authorName}` : ''
-  const postMessage = `This phone number ${phoneNumber} has now been ${action}${byAuthor}.`
-
-  try {
-    await MissiveUtils.createPost(conversationId, postMessage)
-  } catch (error) {
-    console.error(`Failed to create post: ${error}`)
-    throw new Error('Failed to update subscription status and create post.')
-  }
-}
-
 export default {
   makeBroadcast,
-  getAll,
-  patch,
   sendBroadcastFirstMessages,
   sendBroadcastSecondMessages,
   sendNow,
   reconcileTwilioStatus,
-  updateSubscriptionStatus,
   handleFailedDeliveries,
 } as const
