+++ supabase/functions/broadcast-sidebar/index.ts
@@ -1,4 +1,4 @@
-import BroadcastService from '../_shared/services/BroadcastService.ts'
+import BroadcastSidebar from '../_shared/services/BroadcastSidebar.ts'
 import AppResponse from '../_shared/misc/AppResponse.ts'
 import Sentry from '../_shared/lib/Sentry.ts'
 import { Hono } from 'hono'
@@ -9,7 +9,7 @@ app.get('/broadcast-sidebar/', async (c) => {
   const url = new URL(c.req.url)
   const limit = Number(url.searchParams.get('limit')) || 5
   const cursor = Number(url.searchParams.get('cursor')) || undefined
-  const result = await BroadcastService.getAll(limit, cursor)
+  const result = await BroadcastSidebar.getAll(limit, cursor)
   return AppResponse.ok(result)
 })
 
@@ -20,13 +20,26 @@ app.patch('/broadcast-sidebar/', async (c) => {
   }
 
   try {
-    const result = await BroadcastService.patch(Number(id), { firstMessage, secondMessage, runAt, delay })
+    const result = await BroadcastSidebar.patch(Number(id), { firstMessage, secondMessage, runAt, delay })
     return AppResponse.ok(result)
   } catch (error) {
     console.error(`Error: ${error.message}`)
     Sentry.captureException(error)
     return AppResponse.internalServerError()
   }
+})
+
+const corsHeaders = {
+  'Access-Control-Allow-Origin': 'http://localhost:8000',
+  'Access-Control-Allow-Headers': 'authorization, x-client-info, apikey, content-type',
+  'Access-Control-Allow-Methods': 'POST, OPTIONS', // Add allowed methods
+}
+
+app.options('/broadcast-sidebar/', async (c) => {
+  return new Response(null, {
+    status: 204,
+    headers: corsHeaders,
+  })
 })
 
 Deno.serve(app.fetch)
